+++ changelog.md
@@ -2,6 +2,7 @@
 
 <!-- MarkdownTOC -->
 
+- [?](#)
 - [0.5.1](#051)
 - [0.5.0](#050)
 - [0.4.0](#040)
@@ -12,6 +13,16 @@
 - [0.1.0](#010)
 
 <!-- /MarkdownTOC -->
+
+## ?
+
+Released on `?`.
+
+- `databases`
+    + `simbad` and `tap` - querying for a parameter in SIMBAD also returns its bibliographic code, so now the return type for the following functions is `tuple[Any, str]`:
+        * `getStellarParameter()`
+        * `getStellarParameterFromSimbadByMainID()`
+        * `getStellarParameterFromSimbadByObjectID()`
 
 ## 0.5.1